♻️ refactor(components): divide directory recommend screen components
컴포넌트의 파일 명이 길어지는 문제를 해결하고자, 디렉토리를 분리함
