--- src/components/FrameRecommendOrganism/index.tsx
@@ -1,3 +1,0 @@
-import FrameRecommendOrganism from './FrameRecommendOrganism';
-
-export default FrameRecommendOrganism;
