--- src/components/PoseRecommendOrganism/index.tsx
@@ -1,3 +1,0 @@
-import PoseRecommendOrganism from './PoseRecommendOrganism';
-
-export default PoseRecommendOrganism;
