+++ src/components/Recommend/FeedCard/RecommendFeedCard.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {PressableProps} from 'react-native';
 import FastImage from 'react-native-fast-image';
 
-import {PressableLikeIcon} from '../utils/Pressables/PressableIcons';
+import {PressableLikeIcon} from '../../utils/Pressables/PressableIcons';
 import {FeedCardContainer, IconContainer, styles} from './RecommendFeedCard.styles';
 
 type Props = {imgUrl: string} & PressableProps;
