+++ src/components/Recommend/FeedCard/index.tsx
@@ -1,0 +1,3 @@
+import FeedCard from './RecommendFeedCard';
+
+export default FeedCard;
