+++ src/components/Recommend/FrameOrganism/index.tsx
@@ -1,0 +1,3 @@
+import FrameOrganism from './FrameOrganism';
+
+export default FrameOrganism;
