+++ src/components/Recommend/PoseOrganism/index.tsx
@@ -1,0 +1,3 @@
+import PoseOrganism from './PoseRecommendOrganism';
+
+export default PoseOrganism;
