+++ src/components/Recommend/PreviewFourCard/index.tsx
@@ -1,0 +1,3 @@
+import PreviewFourCard from './RecommendPreviewFourCard';
+
+export default PreviewFourCard;
