+++ src/components/Recommend/SituationPoseOrganism/index.tsx
@@ -1,0 +1,3 @@
+import SituationPoseOrganism from './RecommendSituationPoseOrganism';
+
+export default SituationPoseOrganism;
