+++ src/components/Recommend/SituationPoseSwiper/index.tsx
@@ -1,0 +1,3 @@
+import SituationPoseSwiper from './SituationPoseSwiper';
+
+export default SituationPoseSwiper;
