+++ src/components/RecommendDetail/ContentsOrganism/index.tsx
@@ -1,0 +1,3 @@
+import DetailContentsOrganism from './ContentsOrganism';
+
+export default DetailContentsOrganism;
