+++ src/components/RecommendDetail/DiffOrganism/DiffOrganism.styles.tsx
