+++ src/components/RecommendDetail/DiffOrganism/index.tsx
@@ -1,0 +1,3 @@
+import DetailDiffOrganism from './DiffOrganism';
+
+export default DetailDiffOrganism;
