+++ src/components/RecommendDetail/MainFrame/index.tsx
@@ -1,0 +1,3 @@
+import DetailMainFrame from './RecommendDetailMainFrame';
+
+export default DetailMainFrame;
