--- src/components/RecommendDetailContentsOrganism/index.tsx
@@ -1,3 +1,0 @@
-import RecommendDetailContentsOrganism from './RecommendDetailContentsOrganism';
-
-export default RecommendDetailContentsOrganism;
