--- src/components/RecommendDetailDiffOrganism/index.tsx
@@ -1,3 +1,0 @@
-import RecommendDetailDiffOrganism from './RecommendDetailDiffOrganism';
-
-export default RecommendDetailDiffOrganism;
