--- src/components/RecommendDetailMainFrame/index.tsx
@@ -1,3 +1,0 @@
-import RecommendDetailMainFrame from './RecommendDetailMainFrame';
-
-export default RecommendDetailMainFrame;
