--- src/components/RecommendFeedCard/index.tsx
@@ -1,3 +1,0 @@
-import RecommendFeedCard from './RecommendFeedCard';
-
-export default RecommendFeedCard;
