--- src/components/RecommendPreviewFourCard/index.tsx
@@ -1,2 +1,0 @@
-import RecommendPreviewFourCard from './RecommendPreviewFourCard';
-export default RecommendPreviewFourCard;
