--- src/components/RecommendSituationPoseOrganism/index.tsx
@@ -1,2 +1,0 @@
-import RecommendSituationPoseOrganism from './RecommendSituationPoseOrganism';
-export default RecommendSituationPoseOrganism;
