--- src/components/SituationPoseRecommendSwiper/index.tsx
@@ -1,3 +1,0 @@
-import SituationPoseRecommendSwiper from './SituationPoseRecommendSwiper';
-
-export default SituationPoseRecommendSwiper;
