+++ src/screens/RecommendScreen/RecommendDetailScreen.tsx
@@ -2,12 +2,13 @@ import {NativeStackScreenProps} from '@react-navigation/native-stack';
 import React, {useLayoutEffect} from 'react';
 import {SafeAreaView, ScrollView} from 'react-native';
 
+import {DetailHeaderStyle} from './RecommendDetailScreen.header';
+
 import {RecommendParamList} from '.';
 
-import RecommendDetailContentsOrganism from 'src/components/RecommendDetailContentsOrganism';
-import RecommendDetailDiffOrganism from 'src/components/RecommendDetailDiffOrganism';
-import RecommendDetailMainFrame from 'src/components/RecommendDetailMainFrame';
-import {DetailHeaderStyle} from './RecommendDetailScreen.header';
+import RecommendDetailContentsOrganism from 'src/components/RecommendDetail/ContentsOrganism';
+import RecommendDetailDiffOrganism from 'src/components/RecommendDetail/DiffOrganism';
+import RecommendDetailMainFrame from 'src/components/RecommendDetail/MainFrame';
 
 export type DetailScreenProps = NativeStackScreenProps<RecommendParamList, 'RecommendDetail'>;
